+++ src/components/Modal/Modal.module.css
@@ -1,0 +1,18 @@
+.x_axis {
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+.y_axis {
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+
+.modalButton {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 5px;
+  gap: 10px;
+}
