+++ src/components/Modal/Modal.tsx
@@ -1,0 +1,104 @@
+import styles from "./Modal.module.css";
+import { InputText } from "primereact/inputtext";
+import { useRef, useState } from "react";
+import { useAppDispatch, useAppSelector } from "../../store/ChartStore";
+import { addData, openModal } from "../../store/ChartSlice";
+import { Button } from "primereact/button";
+interface DataProps {
+  stateX: string;
+  stateY: string;
+}
+import { Toast } from "primereact/toast";
+import { Dialog } from "primereact/dialog";
+interface ModalProps {
+  handleModal: (value: boolean) => void;
+}
+export default function Modal({ handleModal }: ModalProps) {
+  const [stateX, setStateX] = useState<string>("");
+  const [stateY, setStateY] = useState<string>("");
+  const chartData = useAppSelector((state) => state.chart);
+  const dispatch = useAppDispatch();
+  const toastTL = useRef<Toast>(null);
+
+  function handleSubmit(data: DataProps) {
+    const dataX = data.stateX
+      .split(",")
+      .map((item) => item.trim())
+      .filter((item) => item !== "");
+    const dataY = data.stateY
+      .split(",")
+      .map((item) => item.trim())
+      .filter((item) => item !== "")
+      .map((item) => Number(item))
+      .filter((item) => !isNaN(item));
+
+    if (dataX.length === dataY.length && dataX.length !== 0) {
+      setStateX("");
+      setStateY("");
+      dispatch(
+        addData({
+          x_axis: dataX,
+          y_axis: dataY,
+        })
+      );
+      if (toastTL.current) {
+        toastTL.current.show({
+          severity: "success",
+          summary: "Success",
+          detail: "Chart data updated successfully!",
+          life: 3000,
+        });
+      }
+      setTimeout(() => {
+        handleModal(false);
+      }, 500);
+    } else {
+      if (toastTL.current) {
+        toastTL.current.show({
+          severity: "warn",
+          summary: "Warn",
+          detail: "Add Valid Data!",
+          life: 500,
+        });
+      }
+    }
+  }
+
+  return (
+    <Dialog
+      header="Enter Data"
+      visible={chartData.modalVisible}
+      style={{ width: "50vw" }}
+      onHide={() => {
+        if (!chartData.modalVisible) return;
+        dispatch(
+          openModal({
+            modalVisible: false,
+          })
+        );
+      }}
+    >
+      <div className={styles.x_axis}>
+        <p>X-axis Data (Labels)</p>
+        <span>Enter comma-seperated values</span>
+        <InputText onChange={(e) => setStateX(e.target.value)} value={stateX} />
+      </div>
+
+      <div className={styles.y_axis}>
+        <p>Y-axis Data (Values)</p>
+        <span>Enter comma-seperated values</span>
+        <InputText value={stateY} onChange={(e) => setStateY(e.target.value)} />
+      </div>
+
+      <Toast ref={toastTL} position="top-center" />
+      <div className={styles.modalButton}>
+        <Button
+          type="button"
+          label="Submit"
+          className={styles.buttonchart}
+          onClick={() => handleSubmit({ stateX, stateY })}
+        />
+      </div>
+    </Dialog>
+  );
+}
