+++ src/store/ChartSlice.tsx
@@ -3,6 +3,9 @@ interface addDataPlayload {
   x_axis: Array<string>;
   y_axis: Array<number>;
 }
+interface openModalPlayload {
+  modalVisible: boolean;
+}
 const chartSlice = createSlice({
   name: "Chart",
   initialState: {
@@ -19,6 +22,7 @@ const chartSlice = createSlice({
       "China",
       "Germany",
     ],
+    modalVisible: false,
   },
 
   reducers: {
@@ -26,7 +30,10 @@ const chartSlice = createSlice({
       state.x_axis = action.payload.x_axis;
       state.y_axis = action.payload.y_axis;
     },
+    openModal(state, action: PayloadAction<openModalPlayload>) {
+      state.modalVisible = action.payload.modalVisible;
+    },
   },
 });
-export const { addData } = chartSlice.actions;
+export const { addData, openModal } = chartSlice.actions;
 export default chartSlice.reducer;
